(dev) - config as separate script.
