+++ src/controllers/telegram/common_commands.py
@@ -5,10 +5,11 @@ import logging as log
 import utils
 
 # DEFECT: code duplicate in other scripts
-from bot import bot_config, main_keyboard
+from config import bot_config, main_keyboard
 
 res = bot_config.resources
 
+# ---------------------------------------------
 
 def register_common_commnads(dp: Dispatcher) -> None:
   dp.message.register(start, Command(commands=['start']))
