+++ src/main.py
@@ -1,15 +1,4 @@
-import logging as log
-from environs import Env
-
-env = Env()
-env.read_env()
-
-# #NOTE - configure immediately after import is IMPORTANT!
-log.basicConfig(filename='app.log',
-                level=log._nameToLevel[env('LOG_LEVEL')],
-                format='[{asctime}] {levelname:8} {filename}: {lineno} - {name} - {message}',
-                style='{'
-)
+import config
 
 from aiogram import Dispatcher
 from aiogram.fsm.storage.memory import MemoryStorage
@@ -17,6 +6,7 @@ from aiogram.fsm.storage.memory import MemoryStorage
 from bot import bot
 from controllers.telegram import register_commands
 
+# ---------------------------------------------
 
 # Main
 if __name__ == '__main__':
