+++ src/models/section.py
@@ -4,6 +4,7 @@ from typing import List
 import logging as log
 import os
 
+# ---------------------------------------------
 
 @dataclass
 class Section:
